Massive redo of breadth_first_search().
I gave up on using a trait - too much bureaucracy!

I also gave up on having an iterator for the successors. The issue here is that the successors need an owner, and I couldn't find a good way to do that with an iterator. Instead, we generate a Vec of successors, which serves well as the owner.